ingestion-pipeline/pyproject.toml
[project]
name = "ingestion-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=3.0.0",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.0",
    "sqlalchemy>=2.0.46",
]
